Update API_mem_usage.py
Added 3 more columns named hour, minute, second. 
if the timestamp is present in the parameters of request, the exact data row is the output, else if hour or minute or second is given the rows are given as per required. If that particular data doesn't exist in the database, a message saying it doesn't exist is displayed as an output of the request.